+++ inc-options/wp_nav_menu_options.php
@@ -157,9 +157,9 @@ if ( ! function_exists( 'add_action' ) ) {
 				<tr>
 					<th><?php esc_attr_e( 'Your own Nav Menu options', 'adminimize' );
 						echo '<br />';
-						esc_attr_e( 'ID or class', 'adminimize' ); ?></th>
+						esc_attr_e( 'Option name', 'adminimize' ); ?></th>
 					<th><?php echo '<br />';
-						esc_attr_e( 'Option', 'adminimize' ); ?></th>
+						esc_attr_e( 'Selector, ID or class', 'adminimize' ); ?></th>
 				</tr>
 				</thead>